+++ CHANGES.rst
@@ -9,7 +9,7 @@ Changes to be released in next version
  * VoIP: Change hold direction to send-only.
 
 🐛 Bugfix
- * 
+ * MXSession: Fix app that can fail to resume (vector-im/element-ios/issues/4417).
 
 ⚠️ API Changes
  * MXRoomSummary: `lastMessageEvent` property removed for performance reasons (vector-im/element-ios/issues/4360).